+++ tests/test_connection.py
@@ -1,5 +1,7 @@
 from asyncio import StreamWriter
-from unittest import mock, TestCase
+from unittest import mock
+
+import pytest
 
 from framewirc.client import Client
 from framewirc.connection import Connection
@@ -14,12 +16,11 @@ from framewirc.message import ReceivedMessage
 from .utils import BlankClient
 
 
-class TestRequiredFields(TestCase):
+class TestRequiredFields:
     """Test to show that RequiredAttribuesMixin is properly configured."""
     def test_fields(self):
         """Are the correct fields being checked?"""
-        required = ('client', 'host')
-        self.assertCountEqual(Connection.required_attributes, required)
+        assert Connection.required_attributes == ('client', 'host')
 
     def test_uses_required_attributes_mixin(self):
         """Is RequiredAttributesMixin.__init__ actually getting called?"""
@@ -31,9 +32,9 @@ class TestRequiredFields(TestCase):
         mixin_init.assert_called_with(**kwargs)
 
 
-class ConnectionTestCase(TestCase):
+class ConnectionTestCase:
     """Base TestCase for tests that want an instance of Connection."""
-    def setUp(self):
+    def setup_method(self, method):
         self.connection = Connection(
             client=BlankClient(),
             host='example.com',
@@ -58,7 +59,7 @@ class TestHandle(ConnectionTestCase):
         self.connection.client = mock.MagicMock(spec=Client)
         self.connection.handle(b'')
 
-        self.assertFalse(self.connection.client.on_message.called)
+        assert self.connection.client.on_message.called is False
 
     def test_empty_message(self):
         """
@@ -80,27 +81,27 @@ class TestSend(ConnectionTestCase):
 
     def test_not_bytes(self):
         message = 'PRIVMSG meshy :What µŋhandłed µŋicode yoµ ħave!\r\n'
-        with self.assertRaises(MustBeBytes):
+        with pytest.raises(MustBeBytes):
             self.connection.send(message)
-        self.assertFalse(self.connection.writer.write.called)
+        assert self.connection.writer.write.called is False
 
     def test_only_one_line_ending(self):
         message = b'PRIVMSG meshy :Nice \r\nCODE :injection you have there\r\n'
-        with self.assertRaises(StrayLineEnding):
+        with pytest.raises(StrayLineEnding):
             self.connection.send(message)
-        self.assertFalse(self.connection.writer.write.called)
+        assert self.connection.writer.write.called is False
 
     def test_line_ending_at_eol(self):
         message = b'PRIVMSG meshy :Nice line ending you have forgotten there'
-        with self.assertRaises(NoLineEnding):
+        with pytest.raises(NoLineEnding):
             self.connection.send(message)
-        self.assertFalse(self.connection.writer.write.called)
+        assert self.connection.writer.write.called is False
 
     def test_message_too_long(self):
         message = b'FIFTEEN chars :' + 496 * b'a' + b'\r\n'  # 513 chars
-        with self.assertRaises(MessageTooLong):
+        with pytest.raises(MessageTooLong):
             self.connection.send(message)
-        self.assertFalse(self.connection.writer.write.called)
+        assert self.connection.writer.write.called is False
 
     def test_message_just_right(self):
         message = b'FIFTEEN chars :' + 495 * b'a' + b'\r\n'  # 512 chars
@@ -116,4 +117,4 @@ class TestSendBatch(ConnectionTestCase):
         ]
         self.connection.send_batch(messages)
         calls = self.connection.writer.write.mock_calls
-        self.assertEqual(calls, list(map(mock.call, messages)))
+        assert calls == list(map(mock.call, messages))