+++ tests/test_filters.py
@@ -1,18 +1,18 @@
-from unittest import mock, TestCase
+from unittest import mock
 
 from framewirc import filters
 from framewirc.message import ReceivedMessage
 
 
-class TestCommandBlacklist(TestCase):
-    def setUp(self):
+class TestDeny:
+    def setup_method(self, method):
         self.client = object()
         self.handler = mock.Mock()
 
     def test_correct_list(self):
         """Un-blacklisted commands should be allowed."""
         message = ReceivedMessage(b'COMMAND\r\n')
-        wrapped = filters.command_blacklist(['WRONG_COMMAND'])(self.handler)
+        wrapped = filters.deny(['WRONG_COMMAND'])(self.handler)
 
         wrapped(self.client, message)
 
@@ -24,16 +24,16 @@ class TestCommandBlacklist(TestCase):
     def test_incorrect_list(self):
         """Blacklisted commands should not be allowed."""
         message = ReceivedMessage(b'WRONG_COMMAND\r\n')
-        wrapped = filters.command_blacklist(['WRONG_COMMAND'])(self.handler)
+        wrapped = filters.deny(['WRONG_COMMAND'])(self.handler)
 
         wrapped(self.client, message)
 
-        self.assertFalse(self.handler.called)
+        assert self.handler.called is False
 
     def test_correct_item(self):
         """Un-blacklisted commands should be allowed (string blacklist)."""
         message = ReceivedMessage(b'COMMAND\r\n')
-        wrapped = filters.command_blacklist('WRONG_COMMAND')(self.handler)
+        wrapped = filters.deny('WRONG_COMMAND')(self.handler)
 
         wrapped(self.client, message)
 
@@ -45,22 +45,22 @@ class TestCommandBlacklist(TestCase):
     def test_incorrect_item(self):
         """Blacklisted commands should not be allowed (string blacklist)."""
         message = ReceivedMessage(b'WRONG_COMMAND\r\n')
-        wrapped = filters.command_blacklist('WRONG_COMMAND')(self.handler)
+        wrapped = filters.deny('WRONG_COMMAND')(self.handler)
 
         wrapped(self.client, message)
 
-        self.assertFalse(self.handler.called)
+        assert self.handler.called is False
 
 
-class TestCommandWhitelist(TestCase):
-    def setUp(self):
+class TestAllow:
+    def setup_method(self, method):
         self.client = object()
         self.handler = mock.Mock()
 
     def test_correct_list(self):
         """Whitelisted commands should be allowed."""
         message = ReceivedMessage(b'COMMAND\r\n')
-        wrapped = filters.command_whitelist(['COMMAND'])(self.handler)
+        wrapped = filters.allow(['COMMAND'])(self.handler)
 
         wrapped(self.client, message)
 
@@ -72,16 +72,16 @@ class TestCommandWhitelist(TestCase):
     def test_incorrect_list(self):
         """Unlisted commands should not be allowed."""
         message = ReceivedMessage(b'WRONG_COMMAND\r\n')
-        wrapped = filters.command_whitelist(['COMMAND'])(self.handler)
+        wrapped = filters.allow(['COMMAND'])(self.handler)
 
         wrapped(self.client, message)
 
-        self.assertFalse(self.handler.called)
+        assert self.handler.called is False
 
     def test_correct_item(self):
         """Whitelisted commands should be allowed (string whitelist)."""
         message = ReceivedMessage(b'COMMAND\r\n')
-        wrapped = filters.command_whitelist('COMMAND')(self.handler)
+        wrapped = filters.allow('COMMAND')(self.handler)
 
         wrapped(self.client, message)
 
@@ -93,17 +93,17 @@ class TestCommandWhitelist(TestCase):
     def test_incorrect_item(self):
         """Unlisted commands should not be allowed (string whitelist)."""
         message = ReceivedMessage(b'WRONG_COMMAND\r\n')
-        wrapped = filters.command_whitelist('COMMAND')(self.handler)
+        wrapped = filters.allow('COMMAND')(self.handler)
 
         wrapped(self.client, message)
 
-        self.assertFalse(self.handler.called)
+        assert self.handler.called is False
 
     def test_incorrect_subitem(self):
         """Partial commands should not be allowed (string whitelist)."""
         message = ReceivedMessage(b'COMMA\r\n')
-        wrapped = filters.command_whitelist('COMMAND')(self.handler)
+        wrapped = filters.allow('COMMAND')(self.handler)
 
         wrapped(self.client, message)
 
-        self.assertFalse(self.handler.called)
+        assert self.handler.called is False