+++ tests/test_message.py
@@ -1,67 +1,65 @@
 from itertools import product
-from unittest import mock, TestCase
+from unittest import mock
+
+import pytest
 
 from framewirc import exceptions
 from framewirc.message import build_message, make_privmsgs, ReceivedMessage
 
 
-class TestReceivedMessage(TestCase):
-    """Test the ReceivedMessage class."""
-    def build_message(self, prefix, params, suffix):
-        raw_message = b'COMMAND'
-        if prefix:
-            raw_message = b':prefixed-data ' + raw_message
-        if params:
-            raw_message += b' param1 param2'
-        if suffix:
-            raw_message += b' :suffix message'
-
-        return raw_message
-
-    def test_permutations(self):
-        """
-        Make sure that Message attributes are set correctly.
-
-        Checks every combination of a prefix, params, and suffix data.
-        """
-        on_off = True, False
-        for prefix, params, suffix in product(on_off, on_off, on_off):
-            with self.subTest(prefix=prefix, params=params, suffix=suffix):
-                raw_message = self.build_message(prefix, params, suffix)
-
-                message = ReceivedMessage(raw_message)
-
-                expected_prefix = 'prefixed-data' if prefix else ''
-                expected_params = ('param1', 'param2') if params else ()
-                expected_suffix = 'suffix message' if suffix else ''
-
-                self.assertEqual(message.command, 'COMMAND')
-                self.assertEqual(message.prefix, expected_prefix)
-                self.assertEqual(message.params, expected_params)
-                self.assertEqual(message.suffix, expected_suffix)
+on_off = True, False
 
 
-class TestBuildMessage(TestCase):
+@pytest.mark.parametrize('prefix,params,suffix', product(on_off, on_off, on_off))
+def test_received_message(prefix, params, suffix):
+    """
+    Test the ReceivedMessage class.
+
+    Make sure that Message attributes are set correctly. Checks every
+    combination of a prefix, params, and suffix data.
+    """
+    raw_message = b'COMMAND'
+    if prefix:
+        raw_message = b':prefixed-data ' + raw_message
+    if params:
+        raw_message += b' param1 param2'
+    if suffix:
+        raw_message += b' :suffix message'
+    raw_message += b'\r\n'
+
+    message = ReceivedMessage(raw_message)
+
+    expected_prefix = 'prefixed-data' if prefix else ''
+    expected_params = ('param1', 'param2') if params else ()
+    expected_suffix = b'suffix message' if suffix else b''
+
+    assert message.command == 'COMMAND'
+    assert message.prefix == expected_prefix
+    assert message.params == expected_params
+    assert message.suffix == expected_suffix
+
+
+class TestBuildMessage:
     """Make sure that build_message correctly builds bytes objects."""
     def test_basic(self):
         """Simple command only."""
         message = build_message(b'COMMAND')
-        self.assertEqual(message, b'COMMAND\r\n')
+        assert message == b'COMMAND\r\n'
 
     def test_prefix(self):
         """Command with prefix."""
         message = build_message(b'COMMAND', prefix=b'something')
-        self.assertEqual(message, b':something COMMAND\r\n')
+        assert message == b':something COMMAND\r\n'
 
     def test_params(self):
         """Command with params."""
         message = build_message(b'COMMAND', b'param1', b'param2')
-        self.assertEqual(message, b'COMMAND param1 param2\r\n')
+        assert message == b'COMMAND param1 param2\r\n'
 
     def test_suffix(self):
         """Command with suffix."""
         message = build_message(b'COMMAND', suffix=b'suffix ftw!')
-        self.assertEqual(message, b'COMMAND :suffix ftw!\r\n')
+        assert message == b'COMMAND :suffix ftw!\r\n'
 
     def test_all(self):
         """Command with prefix, params, and suffix."""
@@ -73,7 +71,7 @@ class TestBuildMessage(TestCase):
             suffix=b'suffix ftw!',
         )
         expected = b':something COMMAND param1 param2 :suffix ftw!\r\n'
-        self.assertEqual(message, expected)
+        assert message == expected
 
     def test_unicode(self):
         """
@@ -89,44 +87,44 @@ class TestBuildMessage(TestCase):
             suffix='ftẃ!',
         )
         expected = b':m\xce\xbc COMMAND t\xc3\xa9st test :ft\xe1\xba\x83!\r\n'
-        self.assertEqual(message, expected)
+        assert message == expected
 
 
-class TestBuildMessageExceptions(TestCase):
+class TestBuildMessageExceptions:
     def test_linefeed_in_suffix(self):
         """Make sure that the suffix cannot contain a linefeed (\r\n)."""
-        with self.assertRaises(exceptions.StrayLineEnding):
+        with pytest.raises(exceptions.StrayLineEnding):
             build_message('COMMAND', suffix='\r\n')
 
     def test_linefeed_in_params(self):
         """Make sure that the params cannot contain a linefeed (\r\n)."""
-        with self.assertRaises(exceptions.StrayLineEnding):
+        with pytest.raises(exceptions.StrayLineEnding):
             build_message('COMMAND', '\r\n')
 
     def test_linefeed_in_prefix(self):
         """Make sure that the prefix cannot contain a linefeed (\r\n)."""
-        with self.assertRaises(exceptions.StrayLineEnding):
+        with pytest.raises(exceptions.StrayLineEnding):
             build_message('COMMAND', prefix='\r\n')
 
     def test_linefeed_in_command(self):
         """Make sure that the suffix cannot contain a linefeed (\r\n)."""
-        with self.assertRaises(exceptions.StrayLineEnding):
+        with pytest.raises(exceptions.StrayLineEnding):
             build_message('COMMAND\r\n')
 
     def test_message_too_long(self):
         """Make sure that the the message cannot be longer than 512 bytes."""
-        with self.assertRaises(exceptions.MessageTooLong):
+        with pytest.raises(exceptions.MessageTooLong):
             build_message('A' * 511)  # 513 chars when \r\n added.
 
 
-class TestMakePrivMsgs(TestCase):
+class TestMakePrivMsgs:
     """Ensure make_privmsgs correctly constructs PRIVMSG command lists."""
     def test_simple(self):
         expected = [
             b'PRIVMSG meshy :Just a simple message\r\n',
         ]
         messages = make_privmsgs('meshy', 'Just a simple message')
-        self.assertEqual(messages, expected)
+        assert messages == expected
 
     def test_linebreaks(self):
         expected = [
@@ -138,7 +136,7 @@ class TestMakePrivMsgs(TestCase):
             'meshy',
             'A message\rsplit onto lines\nwith varying linebreaks',
         )
-        self.assertEqual(messages, expected)
+        assert messages == expected
 
     def test_long_line(self):
         too_long = (
@@ -173,7 +171,7 @@ class TestMakePrivMsgs(TestCase):
         ]
 
         messages = make_privmsgs('meshy', too_long)
-        self.assertEqual(messages, expected)
+        assert messages == expected
 
     def test_max_length(self):
         """Is the correct max_length calculated?"""